+++ app/modules/products/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, Float, Text, ARRAY
+from sqlalchemy import Column, Integer, String, Boolean, Text, ARRAY
 from app.core.database import Base
 
 
@@ -8,7 +8,7 @@ class ProductModel(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(100), nullable=False)
     description = Column(Text, nullable=True)
-    price = Column(Float, nullable=False)
+    price = Column(Integer, nullable=False)
     image_urls = Column(ARRAY(String), nullable=True)
     is_active = Column(Boolean, default=True, nullable=False)
     display_order = Column(Integer, default=0, nullable=False)
